@@ -9,81 +9,85 @@
   },
   "license": "MIT",
   "scripts": {
-    "start": "nx serve",
-    "build": "nx build",
-    "test": "nx test"
+    "start": "nx serve starship-mayflower-frontend",
+    "build": "nx build starship-mayflower-frontend",
+    "test": "nx test starship-mayflower-frontend",
+    "lint": "nx lint starship-mayflower-frontend",
+    "type-check": "tsc --noEmit"
   },
   "dependencies": {
-    "@emotion/react": "11.5.0",
-    "@emotion/styled": "11.3.0",
-    "@reduxjs/toolkit": "1.6.2",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@reduxjs/toolkit": "^2.5.0",
     "@types/events": "^3.0.0",
     "@types/paper": "^0.12.3",
     "@types/sylvester-es6": "^0.0.1",
     "@types/three": "^0.135.0",
     "@types/uuid": "^8.3.3",
     "@types/ws": "^8.2.2",
-    "core-js": "^3.6.5",
+    "core-js": "^3.40.0",
     "events": "^3.3.0",
     "isomorphic-ws": "^4.0.1",
     "paper": "^0.12.15",
     "pinus": "^1.4.14",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-redux": "7.2.5",
-    "react-router-dom": "6.1.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^6.28.0",
     "reflect-metadata": "^0.1.13",
     "regenerator-runtime": "0.13.7",
     "sylvester-es6": "^0.0.2",
     "three": "^0.135.0",
-    "tslib": "^2.0.0",
+    "tslib": "^2.8.1",
     "uuid": "^8.3.2",
     "ws": "^8.3.0"
   },
   "devDependencies": {
-    "@babel/core": "7.12.13",
-    "@babel/preset-typescript": "7.12.13",
-    "@emotion/babel-plugin": "11.3.0",
-    "@nrwl/cli": "13.2.3",
-    "@nrwl/cypress": "13.3.0",
-    "@nrwl/eslint-plugin-nx": "13.2.3",
-    "@nrwl/jest": "13.2.3",
-    "@nrwl/linter": "13.2.3",
-    "@nrwl/node": "13.3.1",
-    "@nrwl/react": "13.3.0",
-    "@nrwl/storybook": "^13.3.0",
-    "@nrwl/tao": "13.2.3",
-    "@nrwl/web": "13.3.0",
-    "@nrwl/workspace": "13.2.3",
-    "@storybook/addon-essentials": "~6.4.5",
-    "@storybook/builder-webpack5": "~6.4.5",
-    "@storybook/manager-webpack5": "~6.4.5",
-    "@storybook/react": "~6.4.5",
-    "@svgr/webpack": "^5.4.0",
-    "@testing-library/react": "12.1.2",
-    "@testing-library/react-hooks": "7.0.2",
-    "@types/jest": "27.0.2",
-    "@types/node": "14.14.33",
-    "@types/react": "17.0.30",
-    "@types/react-dom": "17.0.9",
-    "@types/react-router-dom": "5.3.1",
-    "@typescript-eslint/eslint-plugin": "~4.33.0",
-    "@typescript-eslint/parser": "~4.33.0",
-    "babel-jest": "27.2.3",
-    "babel-loader": "8.1.0",
-    "cypress": "^8.3.0",
-    "eslint": "7.32.0",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-import": "2.25.2",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.26.1",
-    "eslint-plugin-react-hooks": "4.2.0",
-    "jest": "27.2.3",
-    "prettier": "^2.3.1",
-    "react-test-renderer": "17.0.2",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.4.3",
-    "url-loader": "^3.0.0"
-  }
+    "@babel/core": "^7.26.0",
+    "@babel/preset-typescript": "^7.26.0",
+    "@emotion/babel-plugin": "^11.13.5",
+    "@eslint/eslintrc": "^3.2.0",
+    "@eslint/js": "^9.17.0",
+    "@nx/cypress": "20.3.0",
+    "@nx/eslint": "20.3.0",
+    "@nx/jest": "20.3.0",
+    "@nx/node": "20.3.0",
+    "@nx/react": "20.3.0",
+    "@nx/storybook": "20.3.0",
+    "@nx/web": "20.3.0",
+    "@nx/webpack": "20.3.0",
+    "@nx/workspace": "20.3.0",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.15",
+    "@storybook/addon-essentials": "~8.4.7",
+    "@storybook/react": "~8.4.7",
+    "@storybook/react-webpack5": "~8.4.7",
+    "@svgr/webpack": "^8.1.0",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/react-hooks": "^8.0.1",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.10.0",
+    "@types/react": "^19.0.1",
+    "@types/react-dom": "^19.0.2",
+    "@types/react-router-dom": "^5.3.3",
+    "babel-jest": "^29.7.0",
+    "babel-loader": "^9.2.1",
+    "cypress": "^13.17.0",
+    "eslint": "^9.17.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-cypress": "^4.1.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "jest": "^29.7.0",
+    "nx": "20.3.0",
+    "prettier": "^3.4.2",
+    "react-refresh": "^0.14.2",
+    "react-test-renderer": "^19.0.0",
+    "ts-jest": "^29.2.5",
+    "typescript": "~5.7.2",
+    "typescript-eslint": "^8.18.1",
+    "url-loader": "^4.1.1"
+  },
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
